Split oversized app event handler tests to satisfy repo lint.
This keeps reveal and ship debug handler coverage while moving groups into dedicated test files so app/test/app_event_handler_scope_test.dart stays below the 1000 non-comment line cap required by repo quality gates.
