+++ app/test/app_event_handler_scope_test.dart
@@ -1,10 +1,8 @@
 import 'package:colonizethis_test/test.dart' show suppressLogsForTests;
 import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_app/core/services/app_event_handler_debug_flip_province.dart';
-import 'package:colonizethis_app/core/services/app_event_handler_debug_reveal_province.dart';
 import 'package:colonizethis_app/core/services/app_event_handler_debug_spawn_civilian.dart';
 import 'package:colonizethis_app/core/services/app_event_handler_debug_spawn_regiment.dart';
-import 'package:colonizethis_app/core/services/app_event_handler_debug_spawn_ship.dart';
 import 'package:colonizethis_app/core/services/app_event_handler_debug_stockpile.dart';
 import 'package:colonizethis_app/core/services/app_event_handler_debug_treasury.dart';
 import 'package:colonizethis_app/core/services/app_event_handler_scope.dart';
@@ -481,92 +479,6 @@ void main() {
       );
       expect(result.game, isNull);
       expect(result.message, contains('unsupported regiment type'));
-    });
-  });
-
-  group('applyDebugShipSpawnAtCapitalHomeFleet', () {
-    test('spawns ships into home fleet and advances global ship sequence', () {
-      final game = Game(
-        id: 'g-ship-1',
-        worldState: WorldState(
-          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 1),
-          oldWorld: RegionData(
-            provinces: [
-              Province(id: 'oldWorld|1', regionId: 'oldWorld', ownerId: 'p1'),
-            ],
-          ),
-          newWorld: const RegionData(),
-          fleets: [
-            Fleet(
-              id: 'fleet_p1',
-              ownerId: 'p1',
-              inPortAtProvinceId: 'oldWorld|1',
-              regionId: 'oldWorld',
-              ships: const [ShipInstance(id: 'ship_3', typeId: 'carrack')],
-            ),
-          ],
-          nextShipInstanceSeq: 4,
-        ),
-        players: const [
-          Player(
-            id: 'p1',
-            displayName: 'P1',
-            isHuman: true,
-            capitalProvinceId: 'oldWorld|1',
-          ),
-        ],
-      );
-      const event = SpawnDebugShipAtCapitalHomeFleetEvent(
-        humanPlayerId: 'p1',
-        shipTypeId: kTechIdShipOfTheLine,
-        count: 2,
-      );
-      final result = applyDebugShipSpawnAtCapitalHomeFleet(
-        currentGame: game,
-        event: event,
-      );
-      final next = result.game;
-      expect(next, isNotNull);
-      expect(result.message, contains('Spawned 2 ship_of_the_line'));
-      final fleet = next!.worldState.fleets.singleWhere(
-        (f) => f.id == 'fleet_p1',
-      );
-      expect(fleet.ships.map((s) => s.id), containsAll(['ship_4', 'ship_5']));
-      expect(next.worldState.nextShipInstanceSeq, 6);
-    });
-
-    test('fails when player has no valid home fleet at capital', () {
-      final game = Game(
-        id: 'g-ship-2',
-        worldState: WorldState(
-          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 1),
-          oldWorld: RegionData(
-            provinces: [
-              Province(id: 'oldWorld|1', regionId: 'oldWorld', ownerId: 'p1'),
-            ],
-          ),
-          newWorld: const RegionData(),
-          fleets: const [],
-        ),
-        players: const [
-          Player(
-            id: 'p1',
-            displayName: 'P1',
-            isHuman: true,
-            capitalProvinceId: 'oldWorld|1',
-          ),
-        ],
-      );
-      const event = SpawnDebugShipAtCapitalHomeFleetEvent(
-        humanPlayerId: 'p1',
-        shipTypeId: 'carrack',
-      );
-      final result = applyDebugShipSpawnAtCapitalHomeFleet(
-        currentGame: game,
-        event: event,
-      );
-      expect(result.game, isNull);
-      expect(result.message, contains('no valid home fleet at capital'));
     });
   });
 
@@ -1115,158 +1027,4 @@ void main() {
     });
   });
 
-  group('applyDebugRevealProvince', () {
-    Game _revealBaseGame() {
-      return Game(
-        id: 'g-reveal',
-        worldState: const WorldState(
-          turnState: TurnState(phase: TurnPhase.orders, turnNumber: 2),
-          oldWorld: RegionData(
-            provinces: [
-              Province(
-                id: 'oldWorld|P1',
-                regionId: 'oldWorld',
-                ownerId: 'ai_1',
-                displayName: 'New Bordeaux',
-              ),
-            ],
-          ),
-          newWorld: RegionData(),
-          tileKeysByRegionAndProvince: {
-            'oldWorld': {
-              'oldWorld|P1': ['oldWorld|P1|0|0'],
-              'oldWorld|s1': ['oldWorld|s1|0|0'],
-            },
-          },
-          playerVisibilityByTile: {
-            'human_1': {
-              'oldWorld|P1|0|0': 'unknown',
-              'oldWorld|s1|0|0': 'unknown',
-            },
-          },
-        ),
-        players: const [
-          Player(id: 'human_1', displayName: 'Human', isHuman: true),
-          Player(id: 'ai_1', displayName: 'AI', isHuman: false),
-        ],
-      );
-    }
-
-    test('reveals province by full id for unknown target', () {
-      final game = _revealBaseGame();
-      const event = RevealDebugProvinceEvent(
-        humanPlayerId: 'human_1',
-        target: 'oldWorld|P1',
-        targetIsFullProvinceId: true,
-      );
-      final result = applyDebugRevealProvince(
-        currentGame: game,
-        event: event,
-        combinedTopology: const MapTopology(),
-      );
-      expect(result.game, isNotNull);
-      final vis = result.game!.worldState.playerVisibilityByTile['human_1']!;
-      expect(vis['oldWorld|P1|0|0'], 'fullyVisible');
-    });
-
-    test('reveal by display name returns disambiguation candidates', () {
-      final game = _revealBaseGame().copyWith(
-        worldState: _revealBaseGame().worldState.copyWith(
-          newWorld: const RegionData(
-            provinces: [
-              Province(
-                id: 'newWorld|P7',
-                regionId: 'newWorld',
-                ownerId: 'ai_1',
-                displayName: 'New Bordeaux',
-              ),
-            ],
-          ),
-        ),
-      );
-      const event = RevealDebugProvinceEvent(
-        humanPlayerId: 'human_1',
-        target: 'New Bordeaux',
-        targetIsFullProvinceId: false,
-      );
-      final result = applyDebugRevealProvince(
-        currentGame: game,
-        event: event,
-        combinedTopology: const MapTopology(),
-      );
-      expect(result.game, isNull);
-      expect(result.message, contains('Candidates: newWorld|P7, oldWorld|P1'));
-      expect(result.message, contains('/reveal_province <regionId|localId>'));
-    });
-
-    test('reveals adjacent sea-zone tiles for coastal province', () {
-      final game = _revealBaseGame();
-      const event = RevealDebugProvinceEvent(
-        humanPlayerId: 'human_1',
-        target: 'oldWorld|P1',
-        targetIsFullProvinceId: true,
-      );
-      final topology = const MapTopology(
-        nodes: [
-          TopologyNode(
-            id: 'P1',
-            regionId: 'oldWorld',
-            type: TopologyNodeType.province,
-          ),
-          TopologyNode(
-            id: 's1',
-            regionId: 'oldWorld',
-            type: TopologyNodeType.seaZone,
-          ),
-        ],
-        edges: [TopologyEdge(id1: 'P1', id2: 's1')],
-      );
-      final result = applyDebugRevealProvince(
-        currentGame: game,
-        event: event,
-        combinedTopology: topology,
-      );
-      final vis = result.game!.worldState.playerVisibilityByTile['human_1']!;
-      expect(vis['oldWorld|s1|0|0'], 'fullyVisible');
-    });
-
-    test('returns deterministic no-op when already fully revealed', () {
-      final game = _revealBaseGame().copyWith(
-        worldState: _revealBaseGame().worldState.copyWith(
-          playerVisibilityByTile: {
-            'human_1': {
-              'oldWorld|P1|0|0': 'fullyVisible',
-              'oldWorld|s1|0|0': 'fullyVisible',
-            },
-          },
-        ),
-      );
-      const event = RevealDebugProvinceEvent(
-        humanPlayerId: 'human_1',
-        target: 'oldWorld|P1',
-        targetIsFullProvinceId: true,
-      );
-      final topology = const MapTopology(
-        nodes: [
-          TopologyNode(
-            id: 'P1',
-            regionId: 'oldWorld',
-            type: TopologyNodeType.province,
-          ),
-          TopologyNode(
-            id: 's1',
-            regionId: 'oldWorld',
-            type: TopologyNodeType.seaZone,
-          ),
-        ],
-        edges: [TopologyEdge(id1: 'P1', id2: 's1')],
-      );
-      final result = applyDebugRevealProvince(
-        currentGame: game,
-        event: event,
-        combinedTopology: topology,
-      );
-      expect(result.message, contains('no-op'));
-    });
-  });
 }
